chore: rename files to kebab-case
